+++ vkconfig_core/configurator.h
@@ -34,7 +34,6 @@
 #include "type_executable_mode.h"
 #include "type_configurator_mode.h"
 #include "type_diagnostic_mode.h"
-#include "type_theme_mode.h"
 #include "serialization.h"
 
 enum EnabledUI {
@@ -119,9 +118,6 @@ class Configurator {
     bool GetShowDiagnosticSearch() const;
     void SetShowDiagnosticSearch(bool enabled);
 
-    ThemeMode GetThemeMode() const;
-    void SetThemeMode(ThemeMode mode);
-
     bool ShouldNotify() const;
 
     bool HasActiveSettings() const;
@@ -170,7 +166,6 @@ class Configurator {
     bool show_diagnostic_search = false;
     ExecutableScope executable_scope = EXECUTABLE_ANY;
     std::string selected_global_configuration = "Validation";
-    ThemeMode theme_mode = THEME_MODE_USE_DEVICE;
 };
 
 struct ConfiguratorGuard {